pyproject.toml
[tool.poetry]
name = "python-django-@"
version = "0.1.0"
description = ""
authors = ["LubHR <liubomyrhruzd89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
